Move TableView and TableModel outside of Rank
